+++ test/ensure_helpers_test.exs
@@ -19,7 +19,7 @@ defmodule MikrotikApi.EnsureHelpersTest do
           {:ok, {200, [], ~s([])}}
 
         _ ->
-          assert method == :post
+          assert method == :put
           assert to_string(url) == "http://10.0.0.1:80/rest/ip/address"
 
           assert Enum.any?(headers, fn {k, v} ->
@@ -66,7 +66,7 @@ defmodule MikrotikApi.EnsureHelpersTest do
           {:ok, {200, [], ~s([])}}
 
         _ ->
-          assert method == :post
+          assert method == :put
           assert to_string(url) == "http://10.0.0.1:80/rest/ip/firewall/filter"
 
           assert Enum.any?(headers, fn {k, v} ->