+++ test/firewall_nat_ensure_test.exs
@@ -19,7 +19,7 @@ defmodule MikrotikApi.FirewallNatEnsureTest do
           {:ok, {200, [], ~s([])}}
 
         _ ->
-          assert method == :post
+          assert method == :put
           assert to_string(url) == "http://10.0.0.1:80/rest/ip/firewall/nat"
 
           assert Enum.any?(headers, fn {k, v} ->