+++ test/test_helper.exs
@@ -1,2 +1,2 @@
-ExUnit.start()
+ExUnit.start(exclude: [:integration])
 Code.require_file("support/transport_mock.ex", __DIR__)